# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/captcha/rotate/check": {
            "post": {
                "description": "验证旋转验证码",
                "tags": [
                    "旋转验证码"
                ],
                "summary": "验证旋转验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证码角度",
                        "name": "angle",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/captcha/shape/slide/check": {
            "get": {
                "description": "验证点击形状验证码",
                "tags": [
                    "点击形状验证码"
                ],
                "summary": "验证点击形状验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "点击坐标",
                        "name": "point",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/captcha/text/check": {
            "get": {
                "description": "验证基础文字验证码",
                "tags": [
                    "基础文字验证码"
                ],
                "summary": "验证基础文字验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证码",
                        "name": "captcha",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "验证码key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/captcha/text/get": {
            "get": {
                "description": "生成基础文字验证码",
                "tags": [
                    "基础文字验证码"
                ],
                "summary": "生成基础文字验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证码类型",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/oauth/github/callback": {
            "get": {
                "description": "登录回调函数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Github OAuth"
                ],
                "summary": "登录回调函数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}